Remove `bundle_install` from public API
First because it's only used for testing.

Second because, only the client knows if it needs to run inside
a bundle clean env and in which path.

So we move this method to the test helper and remove it from the
public API. Clients now need to implement that part themselves.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>