Bump: ignore paths that don't exist
This task was crashing in our code base, due to an extremely large `srb tc` error that wasn't parsed correctly.

I can't copy the error trace because it's basically an entire file, but the gist is that it's trying to wrap that whole class in `T.let(KLASS_NAME...end, T.untyped)`. Anyway, the filename is parsing totally wrong. To avoid errors I'm just ignoring paths that aren't really file paths.