Improve error line matching
By making sure that we can match the whole line as a proper error line,
we can decrease the chances of matching lines that look like error
lines, since they are not indented.

In general, this could be much more expensive to do, however, we already
had the code to match if a line started with an empty space or not. So,
by prefixing the beginning of our error line match to be a
non-whitespace character, we can save on scanning the rest of the string
for a match. Thus, the full regexp match will only execute on
non-indented lines.

The added benefit is that once we match a line to be an error line, we
can match all components of it in a single pass and construct the
`Error` object from those parts.