Fix flaky `test_run_srb_from_bundler_not_found` test
This was actually order dependent.

If this test is executed first and the gem `sorbet` is not installed
on the system Bundler won't know about it and simply exit with a 127
error code.

Whereas if this test is executed _after_ one installing `sorbet`, Bundler
will see the `srb` executable in the GEM_PATH and raise a different error
(explaining that the executable is not in the bundle) and return 1 instead.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>
Co-authored-by: Ufuk Kayserilioglu <ufuk.kayserilioglu@shopify.com>