Do not go through all directory files when bumping
Before this patch, we were listing, reading and checking _all_ the files
(direct and in subdirs) we could find from the bump `directory`.
This may be highly inefficient, especially when the directory contains a
lot of files that will end up ignored from the Sorbet config.

After this patch, while we still list them all, we at least do not read and
check them if they get ignored by the config.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>