support bumping files without typed sigil
Sorbet treats files lacking a typed sigil the same as `typed: false` but
Spoom ignores them. This change identifies such files as `typed: false`
and prepends the `to` sigil to the file as needed.

The complexity of this change is in supporting the "undo" functionality
when spoom backs out a sigil update due to errors or dry run. To
facilitate this, we need to update `change_sigil_in_files` to tell the
caller what the old sigil was so that it can be restored, as it can
now vary by path (typed: false and no sigil now being equivalent).