Do not list files ignored by Sorbet
We ignore files matching the paths ignored in the `sorbet/config` and try to reproduce Sorbet behvior.

From Sorbet docs on `--ignore`:
> Ignores input files that contain the given string in their paths (relative to the input path passed to
> Sorbet). Strings beginning with / match against the prefix of these relative paths; others are substring
> matchs. Matches must be against whole folder and file names, so `foo` matches `/foo/bar.rb` and
> `/bar/foo/baz.rb` but not `/foo.rb` or `/foo2/bar.rb`.

To produce the same behavior, we match files using `File.fnmatch?` and we were prefixing and suffixing the paths with `**`.

When the Sorbet config contains ignored files with extensions like this:

```
--ignore=foo.rb
```

This doesn't work as we are trying to match `foo.rb` with `**/foo.rb/**`.

To fix this problem, we only append `**` to the ignored path IF it does contain a file extension.

Fixes https://github.com/Shopify/spoom/issues/379.

Signed-off-by: Alexandre Terrasa <alexandre.terrasa@shopify.com>