Fix sorbet-static-and-runtime version in tests
Because we do not fix the version of Sorbet used in the Spoom Gemfile,
the Spoom tests have always been run on the newest version of Sorbet.
This can cause tests to fail when a breaking change is made to Sorbet,
without it being obvious to developers why the tests are failing.

This commit fixes the version of `sorbet-static-and-runtime` in the
context of the tests to whatever is currently listed by Gemfile.lock.
Breaking changes in Sorbet will not cause tests to fail until the Sorbet
version is bumped in Gemfile.lock, which leads to test failures that are
easier to identify and fix.