Fix removing first method in class
when following by a multi-line node.

Previously, trying to remove the first method in a module/class when
there was another method after it would also remove all but the last
line of the following method as well. The problem can be seen below:

```
module Foo
  def foo
  end
          # after.location.start_line - 1
  def bar # after.location.end_line - 1
  end
end
```

To account for blank lines, the end_line of the node to remove needs to
be adjusted to include all lines upto the _start_ of the next node,
whereas the code was previously including all lines upto the _end_ of
the next node.