Unset the `DYLD_LIBRARY_PATH` env var when running Sorbet
There are cases where a dev environment can set the
`DYLD_LIBRARY_PATH` env variable in a way that loads a version of
the C++ standard library that is not compatible with the system
version of the library used by Sorbet. This will cause Sorbet to
generate corrupted JSON if the --print=symbol-table-json flag is
passed. Unsetting this environment variable forces Sorbet to default
to the system version of the C++ standard library, preventing this bug.